extracted job executor to jobexcutor package
