--- pkg/executor/client/common.go
@@ -1,9 +1,0 @@
-package client
-
-import (
-	"time"
-)
-
-const (
-	WatchInterval = time.Second
-)
