+++ pkg/executor/jobexecutor/jobexecutor.go
@@ -1,4 +1,4 @@
-package client
+package jobexecutor
 
 import (
 	"bufio"
